+++ docs/source/index.js
@@ -32,3 +32,4 @@ exports['docs/compatibility.pug'] = {
   guide: true
 };
 exports['docs/search.pug'] = { title: 'Search' };
+exports['docs/enterprise.pug'] = { title: 'Mongoose for Enterprise' };