+++ test/index.test.js
@@ -103,12 +103,12 @@ describe('mongoose module:', function() {
   it('{g,s}etting options', function(done) {
     const mongoose = new Mongoose();
 
-    mongoose.set('a', 'b');
-    mongoose.set('long option', 'c');
+    mongoose.set('runValidators', 'b');
+    mongoose.set('useNewUrlParser', 'c');
 
-    assert.equal(mongoose.get('a'), 'b');
-    assert.equal(mongoose.set('a'), 'b');
-    assert.equal(mongoose.get('long option'), 'c');
+    assert.equal(mongoose.get('runValidators'), 'b');
+    assert.equal(mongoose.set('runValidators'), 'b');
+    assert.equal(mongoose.get('useNewUrlParser'), 'c');
     done();
   });
 
@@ -496,6 +496,20 @@ describe('mongoose module:', function() {
     return Promise.resolve();
   });
 
+  it('throws an error on setting invalid options (gh-6899)', function() {
+    let threw = false;
+    try {
+      mongoose.set('someInvalidOption', true);
+    }
+    catch (err) {
+      assert.equal(err.message, '`someInvalidOption` is an invalid option.');
+      threw = true;
+    }
+    finally {
+      assert.equal(threw, true);
+    }
+  });
+
   describe('disconnection of all connections', function() {
     this.timeout(10000);
 
@@ -710,6 +724,12 @@ describe('mongoose module:', function() {
         done();
       });
     });
+  });
+
+  it('isValidObjectId (gh-3823)', function() {
+    assert.ok(mongoose.isValidObjectId('0123456789ab'));
+    assert.ok(mongoose.isValidObjectId(new mongoose.Types.ObjectId()));
+    assert.ok(!mongoose.isValidObjectId(6));
   });
 
   describe('exports', function() {