+++ test/model.indexes.test.js
@@ -7,6 +7,7 @@
 const start = require('./common');
 
 const assert = require('assert');
+const co = require('co');
 const random = require('../lib/utils').random;
 
 const mongoose = start.mongoose;
@@ -104,12 +105,15 @@ describe('model', function() {
       });
     });
 
-    it('of embedded documents unless excludeIndexes (gh-5575)', function(done) {
-      const BlogPost = new Schema({
-        _id: {type: ObjectId},
-        title: {type: String, index: true},
+    it('of embedded documents unless excludeIndexes (gh-5575) (gh-8343)', function(done) {
+      const BlogPost = Schema({
+        _id: { type: ObjectId },
+        title: { type: String, index: true },
         desc: String
       });
+      const otherSchema = Schema({
+        name: { type: String, index: true }
+      }, { excludeIndexes: true });
 
       const User = new Schema({
         name: {type: String, index: true},
@@ -121,7 +125,8 @@ describe('model', function() {
         blogpost: {
           type: BlogPost,
           excludeIndexes: true
-        }
+        },
+        otherArr: [otherSchema]
       });
 
       const UserModel = db.model('gh5575', User);
@@ -283,6 +288,50 @@ describe('model', function() {
       });
     });
 
+    it('when one index creation errors', function(done) {
+      const userSchema = {
+        name: {type: String},
+        secondValue: {type: Boolean}
+      };
+
+      const User = new Schema(userSchema);
+      User.index({ name: 1 });
+
+      const User2 = new Schema(userSchema);
+      User2.index({ name: 1 }, { unique: true });
+      User2.index({ secondValue: 1 });
+
+      const collectionName = 'deepindexedmodel' + random();
+      // Create model with first schema to initialize indexes
+      db.model('SingleIndexedModel', User, collectionName);
+
+      // Create model with second schema in same collection to add new indexes
+      const UserModel2 = db.model('DuplicateIndexedModel', User2, collectionName);
+      let assertions = 0;
+
+      UserModel2.on('index', function() {
+        UserModel2.collection.getIndexes(function(err, indexes) {
+          assert.ifError(err);
+
+          function iter(index) {
+            if (index[0] === 'name') {
+              assertions++;
+            }
+            if (index[0] === 'secondValue') {
+              assertions++;
+            }
+          }
+
+          for (const i in indexes) {
+            indexes[i].forEach(iter);
+          }
+
+          assert.equal(assertions, 2);
+          done();
+        });
+      });
+    });
+
     describe('auto creation', function() {
       it('can be disabled', function(done) {
         const schema = new Schema({name: {type: String, index: true}});
@@ -429,6 +478,32 @@ describe('model', function() {
       return User.init().
         then(() => User.syncIndexes()).
         then(dropped => assert.equal(dropped.length, 0));
+    });
+
+    it('cleanIndexes (gh-6676)', function() {
+      return co(function*() {
+        let M = db.model('gh6676', new Schema({
+          name: { type: String, index: true }
+        }, { autoIndex: false }), 'gh6676');
+
+        yield M.createIndexes();
+
+        let indexes = yield M.listIndexes();
+        assert.deepEqual(indexes.map(i => i.key), [
+          { _id: 1 },
+          { name: 1 }
+        ]);
+
+        M = db.model('gh6676_2', new Schema({
+          name: String
+        }, { autoIndex: false }), 'gh6676');
+
+        yield M.cleanIndexes();
+        indexes = yield M.listIndexes();
+        assert.deepEqual(indexes.map(i => i.key), [
+          { _id: 1 }
+        ]);
+      });
     });
   });
 });